Cleaning and moving props part 11
